make a browser-compatible RemoteFetcher
