+++ package.json
@@ -12,7 +12,7 @@
     "src"
   ],
   "engines": {
-    "node": ">=14.17"
+    "node": ">=18"
   },
   "publishConfig": {
     "access": "public"
